+++ tools/xuoi/shards.cpp
@@ -217,7 +217,7 @@ bool shard_getter(void *data, int idx, const char **out_text)
 {
     auto *items = (std::vector<shard::entry> *)data;
     assert(items);
-    assert(idx < items->size());
+    assert(idx < (int)items->size());
     if (out_text)
         *out_text = items->at(idx).shard_name.c_str();
     return true;
@@ -273,7 +273,7 @@ void write_shards(void *_fp)
         return;
 
     auto fp = (FILE *)_fp;
-    for (int i = 1; i < s_shards.entries.size(); ++i)
+    for (size_t i = 1; i < s_shards.entries.size(); ++i)
     {
         const auto &e = s_shards.entries[i];
         shard::write(fp, e, "Shard");
@@ -290,7 +290,7 @@ int shard_index_by_loginserver(const char *login_server)
 
 shard_data shard_by_id(int id)
 {
-    assert(id < s_shards.entries.size());
+    assert(id < (int)s_shards.entries.size());
     const auto &s = s_shards.entries[id];
     return { s.shard_loginserver.c_str(),
              s.shard_clienttype.c_str(),
